Fix Gemfile to reduce and relax dependencies.
The Gemfile used to list many dependencies with very strict version
requirements, which could conflict with the Rails application this was
used in.

The run-time dependencies have been reduced and given very relaxed version
requirements so that this is as compatible as possible.

The development/test dependencies have been split out into a separate
section so they're only loaded if needed, which avoids having them included
when someone is just trying to use the library.